% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc.st.R
\name{ctmc.st.gth}
\alias{ctmc.st.gth}
\title{GTH algorithm for the stationary solution of continuous-time Markov chain (CTMC)}
\usage{
ctmc.st.gth(Q, nmax = 500)
}
\arguments{
\item{Q}{A n-by-n matrix. A CTMC kernel. matrix, dgeMatrix, dgRMatrix, dgCMatrix and dgTMatrix are allowed.}

\item{nmax}{The maximum size of the matrix Q.}
}
\value{
Stationary vector
}
\description{
Compute the stationary probability vector of CTMC with GTH
\deqn{x Q = 0, \quad x1 = 1}
}
\details{
GTH algorithm is a variation of Guss elimination, and thus this requires n^2 memory.
}
\examples{
Q <- rbind(
  c(-10, 0, 3, 7),
  c(4, -5, 1, 0),
  c(0, 2, -3, 1),
  c(0, 0, 1, -1)
)
ctmc.st.gth(Q)
}
