% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc.tran.R
\name{ctmc.tran.unif}
\alias{ctmc.tran.unif}
\title{Uniformization for transient analysis of countinuous-time Markov chains (CTMCs)}
\usage{
ctmc.tran.unif(Q, x, t, transpose = TRUE,
  eps = sqrt(.Machine$double.eps), ufact = 1.01, rmax = 1000,
  matrix.class = NULL)
}
\arguments{
\item{Q}{A n-by-n matrix. A CTMC kernel. matrix, dgeMatrix, dgRMatrix, dgCMatrix and dgTMatrix are allowed.}

\item{x}{An initial probability vector or a matrix for initial probability vectors.}

\item{t}{A nemeric vector, indicating the time sequence, which should be a non-increasing.}

\item{transpose}{A logical, indicating the matrix Q is transposed or not. The default is TRUE.}

\item{eps}{A numeric value of tolarence error in both uniformization, which determines the rightboud of Poisson distribution.}

\item{ufact}{A numeric value for the uniformization. The default is 1.01.}

\item{rmax}{An integer, indicating the maximum number of rightbound of Poisson distribution in uniformization (see details).}

\item{matrix.class}{A string character, indicating the class of Matrix for the computation.}

\item{r}{A numeric vector or matrix indicating the reward.}
}
\value{
The function returs the following list.
\item{t}{A numeric vector of time points.}
\item{x}{A numeric vector of state probability vector at every time point.}
\item{cx}{A numeric vector of cumulative state probability vector at every time point.}
}
\description{
Compute the state probability vectors and the cumulative values of them at arbitrary time point.
}
\details{
The state probability vector is given by \deqn{x(t) = \exp(Q' t) x} and the cumulative value becomes
\deqn{cx(t) = \int_0^t \exp(Q' u) du x}, where the prime indicates the transpose operator.

\code{ctmc.tran.unif} uses the uniformization to compute the transient solution.

\code{rmax} is the maximum number of rightbound of Poisson distribution. The default is 1000.
The required value depends on the length of time interval and the magnitude of diagonal element of the
infinitesimal generator Q. If the program stops when the required rightbound exceeds \code{rmax},
it can be changed so that \code{rmax} exceeds the requirement value of rightbound.

\code{ctmc.tran.unif} gives the state probability vectors and the cumulative values of them at every time points.
}
\examples{
A <- rbind(
 c(-2,2,0),
 c(3,-5,2),
 c(0,1,-1))
ctmc.tran.unif(Q=A, x=c(1,0,0), t=seq(0,1,length.out=10))

}
