% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.padding.zero.R
\name{diag.padding.zero}
\alias{diag.padding.zero}
\title{Padding diagonal elements}
\usage{
diag.padding.zero(A)
}
\arguments{
\item{A}{A  matrix. This is one of matrix, dgeMatrix, dgCMatrix, dgRMatrix and dgTMatrix}
}
\description{
The function executes the padding diagonal elements as zero if they are skipped in the sparse matrix representation.
}
