% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexp.mix.R
\name{mexpAx.mix}
\alias{mexpAx.mix}
\title{Integral of a matrix exponential with a density function}
\usage{
mexpAx.mix(A, x, f, ..., transpose = FALSE,
  eps = sqrt(.Machine$double.eps), ufact = 1.01, rmax = 1000,
  matrix.class = NULL, inte.func = deformula::deformula.zeroinf)
}
\arguments{
\item{A}{A square matrix.}

\item{x}{A numeric vector or matrix.}

\item{f}{A density function. The parameters of \code{f} are given in the argument \code{...}.}

\item{...}{A list for parameters of \code{f}.}

\item{transpose}{A logical; if TRUE, the matrix A is transposed.}

\item{eps}{A numeric value of tolarence error in uniformization.}

\item{ufact}{A numeric value for the uniformization.}

\item{rmax}{An integer, indicating the maximum number of rightbound of poisson distribution in uniformization.}

\item{matrix.class}{A string character, indicating the class of Matrix for the computation.}

\item{inte.func}{A function that provides the points and weights for numerical integration.}
}
\value{
A vector or matrix of integral of matrix exponential.
}
\description{
Compute a mixture of matrix exponential with a density function.
}
\details{
\code{mexpAx.mix} computes the mixture of matrix exponential of \code{A} with a density function \code{f},
multiplied by a vector or matrix \code{x}, i.e., \eqn{\int \exp(A u) f(u) du x}.
\code{mexpAx.mix} uses the uniformization for the computation of the matrix exponential.
Also, the default numerical integration is the doubly-exponential formulas (DE formulas),
which is provided by the package deformula.

The argument \code{inte.func} provides the points and weights for numerical integration. This function
determines the domain of integration as well as the numerical quadrature. The default is \code{deformula::deformula.zeroinf}
that is the deformula on the domain [0, infinity). The function of \code{inte.func} should returns a list with the elements
\code{x}, \code{w} and \code{h}. The element \code{x} indicates a vector of points for integration which should be a non-decreasing sequence.
\code{w} is a vector of weights corresponding the points. \code{h} is an additional weight. That is, the exact integration should be computed
\eqn{h \sum_i w_i f(x_i)}.
}
\examples{
### integration with Gamma distribution with shape=2 and scale=1 on domain [0, infinity)
A <- rbind(
 c(-2,2,0),
 c(3,-5,2),
 c(0,1,-1))
mexpAx.mix(A=A, f=dgamma, shape=2, scale=1)
}
