% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_matrix.R
\name{special.matrix}
\alias{special.matrix}
\alias{eyeM}
\alias{fillM}
\alias{zeroM}
\alias{vone}
\title{Special sparce matrices}
\usage{
eyeM(n)

fillM(v, m, n = m, dims = c(m, n))

zeroM(m, n = m, dims = c(m, n))

vone(n, byrow = FALSE, elem = 1:n)
}
\arguments{
\item{n}{An integer. The number of rows.}

\item{v}{A numeric value. A value to fill the matrix/vector.}

\item{m}{An integer. The number of rows.}

\item{dims}{A vector for the dimension of matrix.}

\item{byrow}{A logical, indicating whether the vector is a row vector.}

\item{elem}{An integer vector, indicating the elements to be fiiled by 1.}
}
\value{
A special matrix/vector.
}
\description{
Identity matrix, Zero matrix, A matrix whose elements are filled
}
\section{Functions}{
\itemize{
\item \code{eyeM}: An n-by-n identity matrix.

\item \code{fillM}: An m-by-n matrix whose elements are filled by v

\item \code{zeroM}: An m-by-n zero matrix.

\item \code{vone}: A vector whose elements are 1.
}}

